+++ packages/cad-simple-viewer/src/view/AcTrView2d.ts
@@ -29,6 +29,7 @@ import {
   AcEdCalculateSizeCallback,
   AcEdConditionWaiter,
   AcEdCorsorType,
+  AcEdSpatialQueryResultItemEx,
   AcEdViewMode,
   eventBus
 } from '../editor'
@@ -118,7 +119,7 @@ export class AcTrView2d extends AcEdBaseView {
 
     const renderer = new THREE.WebGLRenderer({
       antialias: true,
-      alpha: true,
+      alpha: true
     })
     container.appendChild(renderer.domElement)
 
@@ -411,7 +412,7 @@ export class AcTrView2d extends AcEdBaseView {
    */
   pick(point?: AcGePoint2dLike, hitRadius?: number) {
     if (point == null) point = this.curPos
-    const results: AcDbObjectId[] = []
+    const results: AcEdSpatialQueryResultItemEx[] = []
     const activeLayout = this._scene.activeLayout
     if (activeLayout) {
       const activeLayoutView = this.activeLayoutView
@@ -426,7 +427,7 @@ export class AcTrView2d extends AcEdBaseView {
       firstQueryResults.forEach(item => {
         const objectId = item.id
         if (activeLayout.isIntersectWith(objectId, raycaster)) {
-          results.push(objectId)
+          results.push(item)
         }
       })
     }
@@ -446,7 +447,7 @@ export class AcTrView2d extends AcEdBaseView {
   select(point?: AcGePoint2dLike) {
     const idsAdded: Array<AcDbObjectId> = []
     const results = this.pick(point)
-    results.forEach(id => idsAdded.push(id))
+    results.forEach(item => idsAdded.push(item.id))
     if (idsAdded.length > 0) this.selectionSet.add(idsAdded)
   }
 
@@ -547,6 +548,7 @@ export class AcTrView2d extends AcEdBaseView {
     setTimeout(async () => {
       await this.batchConvert(entities)
     })
+    this._isDirty = true
   }
 
   /**
@@ -555,6 +557,7 @@ export class AcTrView2d extends AcEdBaseView {
   removeEntity(entity: AcDbEntity | AcDbEntity[]) {
     const entities = Array.isArray(entity) ? entity : [entity]
     entities.forEach(entity => this._scene.removeEntity(entity.objectId))
+    this._isDirty = true
   }
 
   /**
@@ -635,11 +638,7 @@ export class AcTrView2d extends AcEdBaseView {
   }
 
   protected createScene() {
-    const scene = new AcTrScene()
-    scene.layouts.forEach(layout => {
-      layout.setSnapObject(this.renderer.createObject())
-    })
-    return scene
+    return new AcTrScene()
   }
 
   private createStats(show?: boolean) {
@@ -817,6 +816,8 @@ export class AcTrView2d extends AcEdBaseView {
       this._scene.addEntity(entity, true)
       entity.dispose()
     })
+    group.dispose()
+
     this._isDirty = true
   }